refactor adding metadata into its own class
