+++ pipeline/beam_tables.py
@@ -19,7 +19,7 @@ import datetime
 import logging
 import pathlib
 import re
-from typing import Optional, Tuple, Dict, List, Any, Iterator, Iterable
+from typing import Any, Dict, List, Optional, Tuple
 
 import apache_beam as beam
 from apache_beam.io.gcp.gcsfilesystem import GCSFileSystem
@@ -27,12 +27,12 @@ from apache_beam.options.pipeline_options import PipelineOptions
 from apache_beam.options.pipeline_options import SetupOptions
 from google.cloud import bigquery as cloud_bigquery  # type: ignore
 
-from pipeline.metadata.beam_metadata import DateIpKey, IP_METADATA_PCOLLECTION_NAME, ROWS_PCOLLECION_NAME, make_date_ip_key, merge_metadata_with_rows
-from pipeline.metadata.schema import BigqueryRow, IpMetadataWithKeys
+from pipeline.metadata.schema import BigqueryRow
+from pipeline.metadata import hyperquack
 from pipeline.metadata import schema
 from pipeline.metadata import flatten_base
-from pipeline.metadata import flatten
 from pipeline.metadata import satellite
+from pipeline.metadata.add_metadata import MetadataAdder
 from pipeline.metadata.ip_metadata_chooser import IpMetadataChooserFactory
 
 # Tables have names like 'echo_scan' and 'http_scan
@@ -248,7 +248,7 @@ class ScanDataBeamPipelineRunner():
     self.bucket = bucket
     self.staging_location = staging_location
     self.temp_location = temp_location
-    self.metadata_chooser_factory = metadata_chooser_factory
+    self.metadata_adder = MetadataAdder(metadata_chooser_factory)
 
   def _get_full_table_name(self, table_name: str) -> str:
     """Get a full project:dataset.table name.
@@ -311,82 +311,6 @@ class ScanDataBeamPipelineRunner():
     ]
 
     return filtered_filenames
-
-  def _add_metadata(
-      self, rows: beam.pvalue.PCollection[BigqueryRow]
-  ) -> beam.pvalue.PCollection[BigqueryRow]:
-    """Add ip metadata to a collection of roundtrip rows.
-
-    Args:
-      rows: beam.PCollection[BigqueryRow]
-
-    Returns:
-      PCollection[BigqueryRow]
-      The same rows as above with with additional metadata columns added.
-    """
-
-    # PCollection[Tuple[DateIpKey,BigqueryRow]]
-    rows_keyed_by_ip_and_date = (
-        rows | 'key by ips and dates' >>
-        beam.Map(lambda row: (make_date_ip_key(row), row)).with_output_types(
-            Tuple[DateIpKey, BigqueryRow]))
-
-    # PCollection[DateIpKey]
-    # pylint: disable=no-value-for-parameter
-    ips_and_dates = (
-        rows_keyed_by_ip_and_date | 'get ip and date keys per row' >>
-        beam.Keys().with_output_types(DateIpKey))
-
-    # PCollection[DateIpKey]
-    deduped_ips_and_dates = (
-        # pylint: disable=no-value-for-parameter
-        ips_and_dates | 'dedup' >> beam.Distinct().with_output_types(DateIpKey))
-
-    # PCollection[Tuple[date,List[ip]]]
-    grouped_ips_by_dates = (
-        deduped_ips_and_dates | 'group by date' >>
-        beam.GroupByKey().with_output_types(Tuple[str, Iterable[str]]))
-
-    # PCollection[Tuple[DateIpKey,IpMetadataWithKeys]]
-    ips_with_metadata = (
-        grouped_ips_by_dates | 'get ip metadata' >> beam.FlatMapTuple(
-            self._add_ip_metadata).with_output_types(Tuple[DateIpKey,
-                                                           IpMetadataWithKeys]))
-
-    # PCollection[Tuple[Tuple[date,ip],Dict[input_name_key,List[BigqueryRow|IpMetadataWithKeys]]]]
-    grouped_metadata_and_rows = (({
-        IP_METADATA_PCOLLECTION_NAME: ips_with_metadata,
-        ROWS_PCOLLECION_NAME: rows_keyed_by_ip_and_date
-    }) | 'group by keys' >> beam.CoGroupByKey())
-
-    # PCollection[BigqueryRow]
-    rows_with_metadata = (
-        grouped_metadata_and_rows |
-        'merge metadata with rows' >> beam.FlatMapTuple(
-            merge_metadata_with_rows).with_output_types(BigqueryRow))
-
-    return rows_with_metadata
-
-  def _add_ip_metadata(
-      self, date: str,
-      ips: List[str]) -> Iterator[Tuple[DateIpKey, IpMetadataWithKeys]]:
-    """Add Autonymous System metadata for ips in the given rows.
-
-    Args:
-      date: a 'YYYY-MM-DD' date key
-      ips: a list of ips
-
-    Yields:
-      Tuples (DateIpKey, IpMetadataWithKeys)
-    """
-    ip_metadata_chooser = self.metadata_chooser_factory.make_chooser(
-        datetime.date.fromisoformat(date))
-
-    for ip in ips:
-      metadata_key = (date, ip)
-      metadata_values = ip_metadata_chooser.get_metadata(ip)
-
-      yield (metadata_key, metadata_values)
 
   def _write_to_bigquery(self, scan_type: str,
                          rows: beam.pvalue.PCollection[BigqueryRow],
@@ -488,18 +412,12 @@ class ScanDataBeamPipelineRunner():
 
       if scan_type == schema.SCAN_TYPE_SATELLITE:
         # PCollection[SatelliteRow]
-        rows = satellite.process_satellite_lines(lines)
+        rows = satellite.process_satellite_lines(lines, self.metadata_adder)
 
       else:  # Hyperquack scans
         # PCollection[HyperquackRow]
-        rows = (
-            lines | 'flatten json' >> beam.ParDo(
-                flatten.FlattenMeasurement()).with_output_types(BigqueryRow))
+        rows = hyperquack.process_hyperquack_lines(lines, self.metadata_adder)
 
-      # PCollection[HyperquackRow|SatelliteRow]
-      rows_with_metadata = self._add_metadata(rows)
+      _raise_error_if_collection_empty(rows)
 
-      _raise_error_if_collection_empty(rows_with_metadata)
-
-      self._write_to_bigquery(scan_type, rows_with_metadata, table_name,
-                              incremental_load)
+      self._write_to_bigquery(scan_type, rows, table_name, incremental_load)
