+++ pipeline/test_beam_tables.py
@@ -13,18 +13,16 @@
 # limitations under the License.
 """Unit tests for the beam pipeline."""
 
+from __future__ import absolute_import
+
 import datetime
 import unittest
-
-import apache_beam as beam
 
 from apache_beam.testing.test_pipeline import TestPipeline
 import apache_beam.testing.util as beam_test_util
 
-from pipeline.metadata.schema import BigqueryRow, IpMetadata, IpMetadataWithKeys
 from pipeline import beam_tables
 from pipeline.metadata.ip_metadata_chooser import FakeIpMetadataChooserFactory
-from pipeline.metadata import satellite
 
 
 class PipelineMainTest(unittest.TestCase):
@@ -32,8 +30,6 @@ class PipelineMainTest(unittest.TestCase):
 
   def setUp(self) -> None:
     self.maxDiff = None  # pylint: disable=invalid-name
-
-  # pylint: disable=protected-access
 
   def test_get_table_name(self) -> None:
     """Test creating a table name given params."""
@@ -69,6 +65,8 @@ class PipelineMainTest(unittest.TestCase):
     self.assertEqual(
         beam_tables.get_job_name('laplante.scan_https', True),
         'append-laplante-scan-https')
+
+  # pylint: disable=protected-access
 
   def test_get_full_table_name(self) -> None:
     project = 'firehook-censoredplanet'
@@ -124,163 +122,6 @@ class PipelineMainTest(unittest.TestCase):
         beam_tables._between_dates(filename, None, datetime.date(2020, 5, 10)))
     self.assertFalse(
         beam_tables._between_dates(filename, datetime.date(2020, 5, 12), None))
-
-  def test_add_metadata(self) -> None:  # pylint: disable=no-self-use
-    """Test adding IP metadata to mesurements."""
-    rows = [
-        BigqueryRow(
-            domain='www.example.com',
-            ip='8.8.8.8',
-            date='2020-01-01',
-            success=True,
-        ),
-        BigqueryRow(
-            domain='www.example.com',
-            ip='1.1.1.1',
-            date='2020-01-01',
-            success=False,
-        ),
-        BigqueryRow(
-            domain='www.example.com',
-            ip='8.8.8.8',
-            date='2020-01-02',
-            success=False,
-        ),
-        BigqueryRow(
-            domain='www.example.com',
-            ip='1.1.1.1',
-            date='2020-01-02',
-            success=True,
-        )
-    ]
-
-    p = TestPipeline()
-    rows = (p | beam.Create(rows))
-
-    runner = beam_tables.ScanDataBeamPipelineRunner(
-        '', '', '', '', FakeIpMetadataChooserFactory())
-
-    rows_with_metadata = runner._add_metadata(rows)
-
-    expected = [
-        BigqueryRow(
-            domain='www.example.com',
-            ip='8.8.8.8',
-            date='2020-01-01',
-            success=True,
-            ip_metadata=IpMetadata(
-                netblock='8.8.8.0/24',
-                asn=15169,
-                as_name='GOOGLE',
-                as_full_name='Google LLC',
-                as_class='Content',
-                country='US',
-            )),
-        BigqueryRow(
-            domain='www.example.com',
-            ip='1.1.1.1',
-            date='2020-01-01',
-            success=False,
-            ip_metadata=IpMetadata(
-                netblock='1.0.0.1/24',
-                asn=13335,
-                as_name='CLOUDFLARENET',
-                as_full_name='Cloudflare Inc.',
-                as_class='Content',
-                country='US',
-            )),
-        BigqueryRow(
-            domain='www.example.com',
-            ip='8.8.8.8',
-            date='2020-01-02',
-            success=False,
-            ip_metadata=IpMetadata(
-                netblock='8.8.8.0/24',
-                asn=15169,
-                as_name='GOOGLE',
-                as_full_name='Google LLC',
-                as_class='Content',
-                country='US',
-            )),
-        BigqueryRow(
-            domain='www.example.com',
-            ip='1.1.1.1',
-            date='2020-01-02',
-            success=True,
-            ip_metadata=IpMetadata(
-                netblock='1.0.0.1/24',
-                asn=13335,
-                as_name='CLOUDFLARENET',
-                as_full_name='Cloudflare Inc.',
-                as_class='Content',
-                country='US',
-            ))
-    ]
-
-    beam_test_util.assert_that(rows_with_metadata,
-                               beam_test_util.equal_to(expected))
-
-  def test_add_ip_metadata_caida(self) -> None:
-    """Test merging given IP metadata with given measurements."""
-    runner = beam_tables.ScanDataBeamPipelineRunner(
-        '', '', '', '', FakeIpMetadataChooserFactory())
-
-    metadatas = list(
-        runner._add_ip_metadata('2020-01-01', ['1.1.1.1', '8.8.8.8']))
-
-    expected_key_1: satellite.DateIpKey = ('2020-01-01', '1.1.1.1')
-    expected_value_1 = IpMetadataWithKeys(
-        ip='1.1.1.1',
-        date='2020-01-01',
-        netblock='1.0.0.1/24',
-        asn=13335,
-        as_name='CLOUDFLARENET',
-        as_full_name='Cloudflare Inc.',
-        as_class='Content',
-        country='US',
-        organization='Fake Cloudflare Sub-Org',
-    )
-
-    expected_key_2: satellite.DateIpKey = ('2020-01-01', '8.8.8.8')
-    expected_value_2 = IpMetadataWithKeys(
-        ip='8.8.8.8',
-        date='2020-01-01',
-        netblock='8.8.8.0/24',
-        asn=15169,
-        as_name='GOOGLE',
-        as_full_name='Google LLC',
-        as_class='Content',
-        country='US',
-        # No organization data is added since the ASN doesn't match dbip
-    )
-
-    self.assertListEqual(metadatas, [(expected_key_1, expected_value_1),
-                                     (expected_key_2, expected_value_2)])
-
-  def disabled_test_add_ip_metadata_maxmind(self) -> None:
-    """Test merging given IP metadata with given measurements."""
-    # TODO turn back on once maxmind is reenabled.
-
-    runner = beam_tables.ScanDataBeamPipelineRunner(
-        '', '', '', '', FakeIpMetadataChooserFactory())
-
-    metadatas = list(runner._add_ip_metadata('2020-01-01', ['1.1.1.3']))
-
-    # Test Maxmind lookup when country data is missing
-    # Cloudflare IPs return Australia
-    expected_key_1 = ('2020-01-01', '1.1.1.3')
-    expected_value_1 = IpMetadataWithKeys(
-        ip='1.1.1.3',
-        date='2020-01-01',
-        netblock='1.0.0.1/24',
-        asn=13335,
-        as_name='CLOUDFLARENET',
-        as_full_name='Cloudflare Inc.',
-        as_class='Content',
-        organization='Fake Cloudflare Sub-Org',
-        country='AU')
-
-    self.assertListEqual(metadatas, [(expected_key_1, expected_value_1)])
 
   # pylint: enable=protected-access
